FIX test_convert_pdf2img.py
