+++ tests/test_convert_pdf2img.py
@@ -1,70 +1,49 @@
 import os
 import unittest
-from typing import List
+from typing import List, Tuple
 
 from parameterized import parameterized
 
+from pdfconduit import Pdfconduit
 from pdfconduit.convert import PDF2IMG
 from pdfconduit.convert.pdf2img import ImageExtension
 from tests import *
 
 
-def convert_pdf2img_params() -> List[str]:
-    return list(
-        map(
-            lambda filename: test_data_path(filename),
-            [
-                "article.pdf",
-                "charts.pdf",
-                "document.pdf",
-                "plan_p.pdf",
-            ],
-        )
+def params() -> List[Tuple[str, ImageExtension, bool]]:
+    files = [
+        "article.pdf",
+        "charts.pdf",
+        "document.pdf",
+        "plan_p.pdf",
+    ]
+    extensions = [ImageExtension.PNG, ImageExtension.JPG]
+    from_stream = [True, False]
+
+    return [
+        (test_data_path(file), extension, use_stream)
+        for use_stream in from_stream
+        for extension in extensions
+        for file in files
+    ]
+
+
+def name_func(testcase_func, param_num, param):
+    return "{}.{}.{}.{}".format(
+        testcase_func.__name__,  # test func
+        param.args[1].value,  # image extension
+        "from_stream" if param.args[2] is True else "from_file",  # stream vs. file
+        get_clean_pdf_name(param.args[0]),  # filename
     )
 
 
-def convert_pdf2img_name_func(testcase_func, param_num, param):
-    return "{}.{}".format(testcase_func.__name__, get_clean_pdf_name(param.args[0]))
-
-
 class TestPdf2Img(PdfconduitTestCase):
-    @parameterized.expand(convert_pdf2img_params, name_func=convert_pdf2img_name_func)
-    def test_convert_pdf_to_images(self, pdf: str):
-        images = PDF2IMG(pdf).convert()
-
-        for image in images:
-            # Assert img file exists
-            self.assertTrue(os.path.exists(image))
-
-            # Assert img file is correct file type
-            self.assertTrue(image.endswith(".png"))
-
-    @parameterized.expand(convert_pdf2img_params, name_func=convert_pdf2img_name_func)
-    def test_convert_pdf_to_images_jpg(self, pdf: str):
-        images = PDF2IMG(pdf, ext=ImageExtension.JPG).convert()
-
-        for image in images:
-            # Assert img file exists
-            self.assertTrue(os.path.exists(image))
-
-            # Assert img file is correct file type
-            self.assertTrue(image.endswith(".jpg"))
-
-    @parameterized.expand(convert_pdf2img_params, name_func=convert_pdf2img_name_func)
-    def test_convert_pdf_to_images_from_stream(self, pdf: str):
-        images = PDF2IMG(self._get_pdf_byte_stream(pdf)).convert()
-
-        for image in images:
-            # Assert img file exists
-            self.assertTrue(os.path.exists(image))
-
-            # Assert img file is correct file type
-            self.assertTrue(image.endswith(".png"))
-
-    @parameterized.expand(convert_pdf2img_params, name_func=convert_pdf2img_name_func)
-    def test_convert_pdf_to_images_jpg_from_stream(self, pdf: str):
+    @parameterized.expand(params, name_func=name_func)
+    def test_convert_pdf_to_images_pdf2img(
+        self, pdf: str, ext: ImageExtension, from_stream: bool
+    ):
         images = PDF2IMG(
-            self._get_pdf_byte_stream(pdf), ext=ImageExtension.JPG
+            pdf if not from_stream else self._get_pdf_byte_stream(pdf), ext=ext
         ).convert()
 
         for image in images:
@@ -72,7 +51,22 @@ class TestPdf2Img(PdfconduitTestCase):
             self.assertTrue(os.path.exists(image))
 
             # Assert img file is correct file type
-            self.assertTrue(image.endswith(".jpg"))
+            self.assertTrue(image.endswith(ext.value))
+
+    @parameterized.expand(params, name_func=name_func)
+    def test_convert_pdf_to_images_pdfconduit(
+        self, pdf: str, ext: ImageExtension, from_stream: bool
+    ):
+        images = Pdfconduit(
+            pdf if not from_stream else self._get_pdf_byte_stream(pdf)
+        ).to_images(ext=ext)
+
+        for image in images:
+            # Assert img file exists
+            self.assertTrue(os.path.exists(image))
+
+            # Assert img file is correct file type
+            self.assertTrue(image.endswith(ext.value))
 
 
 if __name__ == "__main__":
